+++ lib/settings/pages/general_settings_page.dart
@@ -7,6 +7,7 @@ import 'package:lemmy_api_client/v3.dart';
 
 import 'package:thunder/core/enums/local_settings.dart';
 import 'package:thunder/core/enums/nested_comment_indicator.dart';
+import 'package:thunder/core/singletons/lemmy_client.dart';
 import 'package:thunder/core/singletons/preferences.dart';
 import 'package:thunder/settings/widgets/list_option.dart';
 import 'package:thunder/settings/widgets/toggle_option.dart';
@@ -405,11 +406,12 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         ListOption(
                           description: LocalSettings.defaultFeedSortType.label,
                           value: ListPickerItem(label: defaultSortType.value, icon: Icons.local_fire_department_rounded, payload: defaultSortType),
-                          options: allSortTypeItems,
+                          options: [...SortPicker.getDefaultSortTypeItems(includeVersionSpecificFeature: IncludeVersionSpecificFeature.never), ...topSortTypeItems],
                           icon: Icons.sort_rounded,
                           onChanged: (_) {},
                           isBottomModalScrollControlled: true,
                           customListPicker: SortPicker(
+                            includeVersionSpecificFeature: IncludeVersionSpecificFeature.never,
                             title: LocalSettings.defaultFeedSortType.label,
                             onSelect: (value) {
                               setPreferences(LocalSettings.defaultFeedSortType, value.payload.name);
@@ -616,10 +618,11 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         ListOption(
                           description: LocalSettings.defaultCommentSortType.label,
                           value: ListPickerItem(label: defaultCommentSortType.value, icon: Icons.local_fire_department_rounded, payload: defaultCommentSortType),
-                          options: commentSortTypeItems,
+                          options: CommentSortPicker.getCommentSortTypeItems(includeVersionSpecificFeature: IncludeVersionSpecificFeature.never),
                           icon: Icons.comment_bank_rounded,
                           onChanged: (_) {},
                           customListPicker: CommentSortPicker(
+                            includeVersionSpecificFeature: IncludeVersionSpecificFeature.never,
                             title: 'Comment Sort Type',
                             onSelect: (value) {
                               setPreferences(LocalSettings.defaultCommentSortType, value.payload.name);
@@ -628,10 +631,16 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                           ),
                           valueDisplay: Row(
                             children: [
-                              Icon(commentSortTypeItems.firstWhere((sortTypeItem) => sortTypeItem.payload == defaultCommentSortType).icon, size: 13),
+                              Icon(
+                                  CommentSortPicker.getCommentSortTypeItems(includeVersionSpecificFeature: IncludeVersionSpecificFeature.always)
+                                      .firstWhere((sortTypeItem) => sortTypeItem.payload == defaultCommentSortType)
+                                      .icon,
+                                  size: 13),
                               const SizedBox(width: 4),
                               Text(
-                                commentSortTypeItems.firstWhere((sortTypeItem) => sortTypeItem.payload == defaultCommentSortType).label,
+                                CommentSortPicker.getCommentSortTypeItems(includeVersionSpecificFeature: IncludeVersionSpecificFeature.always)
+                                    .firstWhere((sortTypeItem) => sortTypeItem.payload == defaultCommentSortType)
+                                    .label,
                                 style: theme.textTheme.titleSmall,
                               ),
                             ],