+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added Read All in inbox. - contribution from @ggriffo
 - Added support for lemmy 0.19.x authentication - contribution from @micahmo
 - Added support for accessibility profiles in settings - contribution from @micahmo
+- Support new scaled and controversial sort types - contribution from @micahmo
 
 ### Changed
 - Collapsed comments are easier to expand - contribution from @micahmo